The verifier contains both types PTR_TO_BTF_ID and PTR_TO_BTF_ID_OR_NULL. For all other type pairs PTR_TO_foo and PTR_TO_foo_OR_NULL we follow the convention to use PTR_TO_foo_OR_NULL for pointers that may be null and PTR_TO_foo when the ptr value has been checked to ensure it is _not_ NULL.
For PTR_TO_BTF_ID this is not the case though. It may be still be NULL
even though we have the PTR_TO_BTF_ID type.

Improve the comment here to reflect the current state and change the reg
type string to indicate it may be null.  We should try to avoid this in
future types, but its too much code churn to unwind at this point.

Signed-off-by: John Fastabend <john.fastabend@gmail.com>
---
 include/linux/bpf.h   |    2 +-
 kernel/bpf/verifier.c |    2 +-
 2 files changed, 2 insertions(+), 2 deletions(-)