use valid btf in bpf_program__set_attach_target(prog, 0, ...);
bpf_program__set_attach_target() will always fail with fd=0 (attach to a
kernel symbol) because obj->btf_vmlinux is NULL and there is no way to
set it.

Fix this by using libbpf_find_vmlinux_btf_id()

(on a side note: it is unclear whether btf_vmlinux is meant to be
just temporary storage for use in bpf_object__load_xattr(), or
a property of bpf_object, in which case it could be initialuzed
opportunistically, and properly released in bpf_object__close() ).

Signed-off-by: Luigi Rizzo <lrizzo@google.com>
Acked-by: Andrii Nakryiko <andriin@fb.com>