net: ipa: only clear hardware state if setup has completed
In the setup phase of initialization, GSI firmware gets loaded
and initialized, and the AP command TX and default RX endpoints
get set up.  Until that happens, IPA commands must not be issued
to the hardware.

If the modem crashes, we stop endpoint channels and perform a
number of other activities to clear modem-related IPA state,
including zeroing filter and routing tables (using IPA commands).
This is a bug if setup is not complete.  We should also not be
performing the other cleanup activity in that case either.

Fix this by returning immediately when handling a modem crash if we
haven't completed setup.

Fixes: a646d6ec9098 ("soc: qcom: ipa: modem and microcontroller")
Tested-by: Matthias Kaehlcke <mka@chromium.org>
Signed-off-by: Alex Elder <elder@linaro.org>