xsk: fix clang build error in __xp_alloc
Fix a build error with clang in __xp_alloc().

net/xdp/xsk_buff_pool.c:465:15: error: variable 'xskb' is uninitialized
when used here [-Werror,-Wuninitialized]
                        xp_release(xskb);
                                   ^~~~

This is correctly detected by clang, but not gcc. In fact, the
xp_release() statement should not be there at all in the refactored
code, so just remove it.

Fixes: 94033cd8e73b ("xsk: Optimize for aligned case")
Reported-by: Nathan Chancellor <nathan@kernel.org>
Signed-off-by: Magnus Karlsson <magnus.karlsson@intel.com>